feat:reorganizando componentes
