+++ assets/scss/modal.scss
@@ -5,7 +5,7 @@
 }
 .modal-content{
     border-radius: 17px;   
-    box-shadow: inset 0px 0px 2px 1px rgba(0, 0, 0, 0.8);
+    box-shadow: inset 0px 0px 2px 1px rgba(0, 0, 0, 0.2);
 }
 .offcanvas{
     top: 0;
