+++ assets/scss/styles.scss
@@ -9,6 +9,8 @@ xxl(televisão >= 1400px).*/
 @import 'variables.scss';
 @import 'theme_switch.scss';
 @import 'sidebar.scss';
+@import 'table.scss';
+@import 'button.scss';
 
 $dropdown-link-hover-color: white;
 $list-group-hover-bg: map-get($theme-colors, "warning");
@@ -21,4 +23,5 @@ $modal-backdrop-opacity: 0;
 @import "bootstrap/scss/bootstrap";
 @import 'dropdown.scss';
 @import 'modal.scss';
-@import 'search_bar.scss'
+@import 'search_bar.scss';
+@import 'table_mixin.scss';
