+++ assets/scss/table.scss
@@ -1,0 +1,2 @@
+@import 'variables.scss';
+$table-hover-bg: rgb(254, 213, 30, 0.5);
