+++ components/header/Brand.vue
@@ -1,0 +1,18 @@
+<template>
+    <div class="logo">
+          <a class="navbar-brand" href="/">
+            <svg class="d-flex align-items-center" width="53" height="55">
+              <image class="custom-logo" xlink:href="/logo.svg" width="52" height="52"/>
+            </svg>
+          </a>
+        </div>
+</template>
+
+<style scoped>
+.custom-logo{
+    transition: transform 0.3s ease-in-out;
+}
+.custom-logo:hover{
+  transform: scale(1.07);
+}
+</style>
