+++ components/header/ThemeSwitch.vue
@@ -1,0 +1,5 @@
+<template>
+    <div class="form-check form-switch">
+        <input class="form-check-input shadow-none" type="checkbox" role="switch" id="flexSwitchCheckDefault"> 
+    </div>
+</template>
