--- components/header/search/SearchBar.vue
@@ -1,5 +1,0 @@
-<template>
-    <div>
-      
-    </div>
-</template>
