--- components/header/theme_switch/ThemeSwitch.vue
@@ -1,5 +1,0 @@
-<template>
-    <div>
-      
-    </div>
-</template>
