+++ components/tables/salvo.vue
@@ -1,0 +1,208 @@
+<template>
+    <div>
+      <table class="table table-hover">
+        <thead>
+            <tr>
+                <th class="col-title" scope="col">Nome</th>
+                <th class="col-title" scope="col">Código Sipac</th>
+                <th class="col-title" scope="col">Status do item</th>
+                <th class="col-title" scope="col">Quantidade</th>
+                <th class="col-title end"></th>
+            </tr>
+        </thead>
+        <tbody>
+            <tr>
+                <th class="cel" scope="row"><p>Cartolina Amarela</p></th>
+                <th><p>283492354</p></th>
+                <th><p>Disponível</p></th>
+                <th><p>2</p></th>
+                <th class="end">
+                    <button class="btn btn-primary">
+                        Detalhes
+                    </button>
+                    <button class="btn btn-primary">
+                        Histórico
+                    </button>
+                    <button class="btn btn-primary">
+                        Adicionar
+                    </button>
+                    <button class="btn btn-primary">
+                        Remover
+                    </button>
+                </th>
+            </tr>
+            <tr>
+                <th class="cel" scope="row"><p>Cartolina Amarela</p></th>
+                <th><p>283492354</p></th>
+                <th><p>Disponível</p></th>
+                <th><p>2</p></th>
+                <th class="end">
+                    <button class="btn btn-primary">
+                        Detalhes
+                    </button>
+                    <button class="btn btn-primary">
+                        Histórico
+                    </button>
+                    <button class="btn btn-primary">
+                        Adicionar
+                    </button>
+                    <button class="btn btn-primary">
+                        Remover
+                    </button>
+                </th>
+            </tr>
+            <tr>
+                <th class="cel" scope="row"><p>Cartolina Amarela</p></th>
+                <th><p>283492354</p></th>
+                <th><p>Disponível</p></th>
+                <th><p>2</p></th>
+                <th class="end">
+                    <button class="btn btn-primary">
+                        Detalhes
+                    </button>
+                    <button class="btn btn-primary">
+                        Histórico
+                    </button>
+                    <button class="btn btn-primary">
+                        Adicionar
+                    </button>
+                    <button class="btn btn-primary">
+                        Remover
+                    </button>
+                </th>
+            </tr>
+            <tr>
+                <th class="cel" scope="row"><p>Cartolina Amarela</p></th>
+                <th><p>283492354</p></th>
+                <th><p>Disponível</p></th>
+                <th><p>2</p></th>
+                <th class="end">
+                    <button class="btn btn-primary">
+                        Detalhes
+                    </button>
+                    <button class="btn btn-primary">
+                        Histórico
+                    </button>
+                    <button class="btn btn-primary">
+                        Adicionar
+                    </button>
+                    <button class="btn btn-primary">
+                        Remover
+                    </button>
+                </th>
+            </tr>
+            <tr>
+                <th class="cel" scope="row"><p>Cartolina Amarela</p></th>
+                <th><p>283492354</p></th>
+                <th><p>Disponível</p></th>
+                <th><p>2</p></th>
+                <th class="end">
+                    <button class="btn btn-primary">
+                        Detalhes
+                    </button>
+                    <button class="btn btn-primary">
+                        Histórico
+                    </button>
+                    <button class="btn btn-primary">
+                        Adicionar
+                    </button>
+                    <button class="btn btn-primary">
+                        Remover
+                    </button>
+                </th>
+            </tr>
+            <tr>
+                <th class="cel" scope="row"><p>Cartolina Amarela</p></th>
+                <th><p>283492354</p></th>
+                <th><p>Disponível</p></th>
+                <th><p>2</p></th>
+                <th class="end">
+                    <button class="btn btn-primary">
+                        Detalhes
+                    </button>
+                    <button class="btn btn-primary">
+                        Histórico
+                    </button>
+                    <button class="btn btn-primary">
+                        Adicionar
+                    </button>
+                    <button class="btn btn-primary">
+                        Remover
+                    </button>
+                </th>
+            </tr>
+            <tr>
+                <th class="cel" scope="row"><p>Cartolina Amarela</p></th>
+                <th><p>283492354</p></th>
+                <th><p>Disponível</p></th>
+                <th><p>2</p></th>
+                <th class="end">
+                    <button class="btn btn-primary">
+                        Detalhes
+                    </button>
+                    <button class="btn btn-primary">
+                        Histórico
+                    </button>
+                    <button class="btn btn-primary">
+                        Adicionar
+                    </button>
+                    <button class="btn btn-primary">
+                        Remover
+                    </button>
+                </th>
+            </tr>
+            <tr>
+                <th class="cel" scope="row"><p>Cartolina Amarela</p></th>
+                <th><p>283492354</p></th>
+                <th><p>Disponível</p></th>
+                <th><p>2</p></th>
+                <th class="end">
+                    <button class="btn btn-primary">
+                        Detalhes
+                    </button>
+                    <button class="btn btn-primary">
+                        Histórico
+                    </button>
+                    <button class="btn btn-primary">
+                        Adicionar
+                    </button>
+                    <button class="btn btn-primary">
+                        Remover
+                    </button>
+                </th>
+            </tr>
+        </tbody>
+      </table>
+    </div>
+</template>
+
+<style scoped>
+table{
+    position: relative;
+    left: 0px;
+}
+.col-title{
+    font-size: 15px;
+    font-weight: 500;
+}
+th { 
+    text-decoration: none;
+    text-align: center;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    padding: 20px 5vw 20px 1vw;
+}
+p{
+    margin: 0;
+    
+}
+.btn{
+    opacity: 100%;
+    margin-top: 10px;
+    margin-right: 10px;
+}
+tr:hover .btn{
+    display: inline-block;
+}
+</style>
