+++ pages/inventario.vue
@@ -1,28 +1,5 @@
 <template>
-    <div class="container col-1 d-block" style="margin-left: 170px;">
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
-        <p>Inventário</p>
+    <div class="container d-block" style="margin-left: 10px;">
+        <TablesTable />
     </div>
 </template>
